Added button and compliment selection functionality. Tied it to makeCall and storeCompliment.
